# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super STT daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/stt/super-stt-http.sock` (override with `SUPER_STT_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/stt/super-stt-http.sock\" \\\n     -H \"Authorization: Bearer $STT_TOKEN\" \\\n     http://stt.local/v1/ping\n```",
    "contact": {
      "name": "Super STT",
      "url": "https://github.com/jorge-menjivar/super-stt"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.2.4-beta.1"
  },
  "servers": [
    {
      "url": "http://stt.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/backends/{source}/models/{model}/language": {
      "get": {
        "tags": [
          "backends"
        ],
        "summary": "Read a model's language override",
        "description": "The language this specific model transcribes in, which overrides the global `/language` setting. Addressed by `(source, model)` rather than \"the active model\", so it can be read whether or not the model is loaded.\n\n`null` means no override: the model follows the global setting.",
        "operationId": "get_one",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "description": "The backend's `source`, percent-encoded — e.g. `github.com%2Facme%2Fwhisper`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Facme%2Fwhisper"
          },
          {
            "name": "model",
            "in": "path",
            "description": "The model's name, as `GET /backends` spells it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "How this model's language resolves.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelLanguageState"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such backend (`unknown_backend`) or no such model (`unknown_model`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "backends"
        ],
        "summary": "Set a model's language override",
        "description": "Pins this model to a language regardless of the global `/language` setting. A tag the model does not serve is refused rather than silently ignored.\n\nOverridden in turn by a `language` field in a single `POST /transcribe` body.",
        "operationId": "set",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "description": "The backend's `source`, percent-encoded — e.g. `github.com%2Facme%2Fwhisper`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Facme%2Fwhisper"
          },
          {
            "name": "model",
            "in": "path",
            "description": "The model's name, as `GET /backends` spells it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Override set.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelLanguageState"
                }
              }
            }
          },
          "400": {
            "description": "The body was empty (`invalid_request`), or this model does not serve that language (`unsupported_language`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such backend (`unknown_backend`) or no such model (`unknown_model`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "backends"
        ],
        "summary": "Clear a model's language override",
        "description": "Removes the per-model pin, returning this model to the global `/language` setting.",
        "operationId": "clear",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "description": "The backend's `source`, percent-encoded — e.g. `github.com%2Facme%2Fwhisper`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Facme%2Fwhisper"
          },
          {
            "name": "model",
            "in": "path",
            "description": "The model's name, as `GET /backends` spells it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Override cleared.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelLanguageState"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such backend (`unknown_backend`) or no such model (`unknown_model`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
    "/v1/backends/{source}/options/list": {
      "get": {
        "tags": [
          "backends"
        ],
        "summary": "List a backend's options",
        "description": "Every option this backend declares, each with its label, type, default, and the value actually in effect. This is what a settings UI renders a form from.\n\nOptions are the backend's own configuration — an endpoint URL, a model parameter — declared in its manifest. Credentials are not options; those are secrets, at `/backends/{source}/secrets/list`.",
        "operationId": "list",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "description": "The backend's `source`, percent-encoded — e.g. `github.com%2Facme%2Fwhisper`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Facme%2Fwhisper"
          }
        ],
        "responses": {
          "200": {
            "description": "The backend's options.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackendOptions"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No installed backend has that `source` (`unknown_backend`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/backends/{source}/secrets/list": {
      "get": {
        "tags": [
          "backends"
        ],
        "summary": "List a backend's secrets",
        "description": "Every credential this backend declares, with whether each is currently set. **Values are never returned** — not here, not anywhere. They live in the system keyring, and the only operations are write and clear.\n\nThis is a separate scope from `settings` precisely because it is the credential surface: a token granted `settings` cannot reach it.",
        "operationId": "list",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "description": "The backend's `source`, percent-encoded — e.g. `github.com%2Facme%2Fopenai`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "github.com%2Facme%2Fopenai"
          }
        ],
        "responses": {
          "200": {
            "description": "The declared secrets and whether each is set.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretList"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `secrets` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such backend (`unknown_backend`) or no such secret (`unknown_secret`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "secrets"
            ]
          }
        ]
      }
    },
...
    "/v1/pipeline/{stage}/model/reload": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Re-instantiate a stage's model in place",
        "description": "Tears the model down and brings it back up so it picks up changed secrets and options — an API key set through `/backends/{source}/secrets`, say — without the client having to unload and reload by hand.\n\nNothing is re-downloaded; the files on disk are unchanged.",
        "operationId": "reload_stage_model",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position: `1` transcribes, `2` post-processes. A position that does not exist is a `404 unknown_stage`.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Reloaded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "The error envelope, as [`transport.md`] specifies it.\n\n`error_code` is the stable, machine-readable identifier clients switch on\nand the field that determines the HTTP status; `message` is prose for a\nhuman and may be reworded at any time.\n\n[`transport.md`]: https://github.com/jorge-menjivar/super-stt/blob/main/docs/protocol/transport.md",
        "required": [
          "status"
        ],
        "properties": {
          "error_code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "The stable identifier for this failure. Absent only on an unclassified\nserver-side error."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "LanguageBody": {
        "type": "object",
        "description": "The language this model should transcribe in.",
        "properties": {
          "language": {
            "type": "string",
            "description": "A BCP-47 tag such as `es`, or `auto` to let the model detect it. Empty is\nrefused — clear an override with `DELETE`.",
            "example": "es"
          }
        }
      },
...
      "ModelLanguageState": {
        "type": "object",
        "description": "The per-model language resolution.",
        "required": [
          "status",
          "language"
        ],
        "properties": {
          "language": {
            "$ref": "#/components/schemas/ModelLanguageBlock"
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "Reason": {
        "type": "object",
        "description": "The `data` object of a [`ReasonEnvelope`].",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Which case of `message` occurred — e.g. `expired`, `exe_changed`,\n`user_denied`."
          }
        }
      },
      "ReasonEnvelope": {
        "type": "object",
        "description": "The auth-failure envelope: `message` names the failure and `data.reason`\nsays which of its cases occurred.\n\nDistinct from [`ErrorEnvelope`] because the auth surface predates\n`error_code` and clients read `data.reason` there. Both are documented\nrather than reconciled, since changing either is a breaking wire change.",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reason"
          },
          "message": {
            "type": "string",
            "description": "The failure identifier, e.g. `invalid_session` or `auth_denied`."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}